Move `compiler::Resolved` to `core`
