+++ lib/vrl/core/Cargo.toml
@@ -8,3 +8,4 @@ publish = false
 [dependencies]
 lookup = { path = "../../lookup" }
 value = { path = "../../value" }
+diagnostic = { package = "vrl-diagnostic", path = "../diagnostic" }
