+++ lib/vrl/core/src/lib.rs
@@ -1,5 +1,7 @@
+mod expression;
 mod r#macro;
 mod target;
 
+pub use expression::{ExpressionError, Resolved};
 pub use target::Target;
 pub use value::Value;
